HCNOVEO-1794 Fix compose file list progress
